pyproject.toml
[project]
name = "english-learning"
version = "0.1.0"
description = "English Learning Material Management System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "apscheduler>=3.10.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.5.0",
    "beautifulsoup4>=4.12.3",
    "aiofiles>=24.1.0",
    "email-validator>=2.1.0",
]
